requirements.txt
google-genai>=0.12.0
requests>=2.28
python-dotenv>=1.0.0